DB infastructure change
